feat: raw binary request body + R2 bucket binding + CI fix
- Update tinywasm/router to v0.1.7 and implement PublicAsset/PublicDir.
- Fix binary corruption by using ArrayBuffer and Uint8Array in workers and r2.
- Implement lazy body reading in workers.Request.
- Add r2 package for Cloudflare R2 bucket integration.
- Support R2_BUCKET_ID and R2_BUCKET_NAME in configuration.
- Include D1 and R2 bindings (type r2_bucket) in DeployWorker metadata.
- Add binary round-trip tests for R2.
- Fix CI failure by adding GitHub authentication via .netrc to the test workflow.
- Prevent size truncation in R2 List by using Float().

+++ r2/bucket.go
@@ -110,7 +110,7 @@ func (b *Bucket) List(prefix string) ([]ObjectInfo, error) {
 		o := jsObjects.Index(i)
 		info := ObjectInfo{
 			Key:  o.Get("key").String(),
-			Size: int64(o.Get("size").Int()),
+			Size: int64(o.Get("size").Float()),
 		}
 		httpMetadata := o.Get("httpMetadata")
 		if !httpMetadata.IsUndefined() && !httpMetadata.IsNull() {
